+++ Dataset_netCDF_2.py
@@ -32,57 +32,53 @@ def Ux_it_offset(it):
     return np.average(Ux_rotor)
 
 
-# def IA_it_offset(it):
+def IA_it_offset(it):
+
+    velx = np.reshape(velocityx[it],(y,x)); vely = np.reshape(velocityy[it],(y,x))
+
+    fx = interpolate.interp2d(Y,Z,velx,kind="linear"); fy = interpolate.interp2d(Y,Z,vely,kind="linear")
+
+    IA = 0
+    for j,k in zip(ys,zs):
+        r = np.sqrt(j**2 + k**2)
+        if r <= 63 and r > 1.5:
+            delta_Ux_i = delta_Ux(j,k,r,fx,fy)
+            IA += r * delta_Ux_i * dA
+    return IA
 
 
-#     hvelmag_interp = hvelmag[it]
-#     hvelmag_interp = np.reshape(hvelmag_interp,(y,x))
-#     f = interpolate.interp2d(ys,zs,hvelmag_interp)
+def delta_Ux(j,k,r,fx,fy):
 
-#     Hvelmag = hvelmag[it]
-#     Hvelmag = np.reshape(Hvelmag,(y,x))
+    theta = np.arccos(j/r)
 
-#     IA = 0
-#     for j in np.arange(0,len(ys)):
-#         for k in np.arange(0,len(zs)):
-#             r = np.sqrt(ys[j]**2 + zs[k]**2)
-#             if r <= 63 and r >= 1.5:
+    if theta + ((2*np.pi)/3) > (2*np.pi):
+        theta_1 = theta +(2*np.pi)/3 - (2*np.pi)
+    else:
+        theta_1 = theta + (2*np.pi)/3
 
-#                 delta_Ux_i = delta_Ux(r,j,k,f,Hvelmag)
-#                 IA += r * delta_Ux_i * dA
-#     return IA
+    Y_1 = r*np.cos(theta_1)
+    Z_1 = r*np.sin(theta_1)
 
 
-# def delta_Ux(r,j,k,f,Hvelmag):
+    if theta - ((2*np.pi)/3) < 0:
+        theta_2 = theta - ((2*np.pi)/3) + (2*np.pi)
+    else:
+        theta_2 = theta - ((2*np.pi)/3)
 
-#     Y_0 = ys[j]
+    Y_2 = r*np.cos(theta_2)
+    Z_2 = r*np.sin(theta_2)
 
-#     theta = np.arccos(Y_0/r)
+    vx = fx(j,k); vy = fy(j,k)
+    vx_1 = fx(Y_1,Z_1); vy_1 = fy(Y_1,Z_1)
+    vx_2 = fx(Y_2,Z_2); vy_2 = fy(Y_2,Z_2)
 
-#     if theta + ((2*np.pi)/3) > (2*np.pi):
-#         theta_1 = theta +(2*np.pi)/3 - (2*np.pi)
-#     else:
-#         theta_1 = theta + (2*np.pi)/3
+    Ux_0 = vx*np.cos(np.radians(29))+vy*np.sin(np.radians(29))
+    Ux_1 = vx_1*np.cos(np.radians(29))+vy_1*np.sin(np.radians(29))
+    Ux_2 = vx_2*np.cos(np.radians(29))+vy_2*np.sin(np.radians(29))
 
-#     Y_1 = r*np.cos(theta_1)
-#     Z_1 = r*np.sin(theta_1)
+    delta_Ux =  np.max( [abs( Ux_0 - Ux_1 ), abs( Ux_0 - Ux_2 )] )
 
-
-#     if theta - ((2*np.pi)/3) < 0:
-#         theta_2 = theta - ((2*np.pi)/3) + (2*np.pi)
-#     else:
-#         theta_2 = theta - ((2*np.pi)/3)
-
-#     Y_2 = r*np.cos(theta_2)
-#     Z_2 = r*np.sin(theta_2)
-
-#     Ux_0 =  Hvelmag[j][k]
-#     Ux_1 =  f(Y_1,Z_1)
-#     Ux_2 =  f(Y_2,Z_2)
-
-#     delta_Ux =  np.max( [abs( Ux_0 - Ux_1 ), abs( Ux_0 - Ux_2 )] )
-
-#     return delta_Ux
+    return delta_Ux
 
 
 
@@ -183,20 +179,17 @@ for offset in offsets:
     group = ncfile.createGroup("{}".format(group_label[ic]))
 
     Ux = group.createVariable("Ux", np.float64, ('sampling'),zlib=True)
-    #IA = group.createVariable("IA", np.float64, ('sampling'),zlib=True)
+    IA = group.createVariable("IA", np.float64, ('sampling'),zlib=True)
 
     p_rotor = a.groups["p_r"]; del a
 
     velocityx = np.array(p_rotor.variables["velocityx"]); velocityy = np.array(p_rotor.variables["velocityy"])
 
-    # Variables = ["Ux_{0}".format(offset), "IA_{0}".format(offset)]
-    # units = ["[m/s]", "[$m^4/s$]"]
-    Variables = ["Ux_{0}".format(offset)]
-    units = ["[m/s]"]
+    Variables = ["Ux_{0}".format(offset), "IA_{0}".format(offset)]
+    units = ["[m/s]", "[$m^4/s$]"]
 
     x = p_rotor.ijk_dims[0] #no. data points
     y = p_rotor.ijk_dims[1] #no. data points
-
 
     coordinates = np.array(p_rotor.variables["coordinates"])
 
@@ -217,9 +210,10 @@ for offset in offsets:
     Y = np.linspace(round(np.min(ys),0), round(np.max(ys),0),x )
     Z = np.linspace(round(np.min(zs),0), round(np.max(zs),0),y )
 
+    del coordinates
 
-    dy = ys[1]-ys[0]
-    dz = zs[1] - zs[0]
+    dy = (max(Y) - min(Y))/x
+    dz = (max(Z) - min(Z))/y
     dA = dy * dz
 
     del p_rotor
@@ -242,17 +236,17 @@ for offset in offsets:
                 Ux_it = np.array(Ux_it)
                 Ux[:] = Ux_it; del Ux_it
 
-        # elif Variable[0:2] == "IA":
-        #     IA_it = []
-        #     print("IA calcs")
-        #     with Pool() as pool:
-        #         i_IA = 1
-        #         for IA_i in pool.imap(IA_it_offset, np.arange(0,time_idx)):
-        #             IA_it.append(IA_i)
-        #             print(i_IA,time.time()-start_time)
-        #             i_IA+=1
-        #         IA_it = np.array(IA_it)
-        #         IA[:] = IA_it; del IA_it
+        elif Variable[0:2] == "IA":
+            IA_it = []
+            print("IA calcs")
+            with Pool() as pool:
+                i_IA = 1
+                for IA_i in pool.imap(IA_it_offset, np.arange(0,time_idx)):
+                    IA_it.append(IA_i)
+                    print(i_IA,time.time()-start_time)
+                    i_IA+=1
+                IA_it = np.array(IA_it)
+                IA[:] = IA_it; del IA_it
 
 
     print(ncfile.groups)
